# session-11-crud-user-grpcgateway-cache/docs/proto/user_service/v1/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User Service",
    "description": "User service handle API for CRUD users",
    "version": "v1"
  },
  "tags": [
    {
      "name": "proto.user_service.v1.UserService"
    }
  ],
  "host": "localhost:8080",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/users": {
      "get": {
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.GetUsersResponse"
            }
          }
        },
        "tags": [
          "proto.user_service.v1.UserService"
        ]
      },
      "post": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.MutationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.CreateUserRequest"
            }
          }
        ],
        "tags": [
          "proto.user_service.v1.UserService"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "GetUserByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.GetUserByIDResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "proto.user_service.v1.UserService"
        ]
      },
      "delete": {
        "operationId": "DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.MutationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.UserService.DeleteUserBody"
            }
          }
        ],
        "tags": [
          "proto.user_service.v1.UserService"
        ]
      },
      "put": {
        "operationId": "UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.MutationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/proto.user_service.v1.UserService.UpdateUserBody"
            }
          }
        ],
        "tags": [
          "proto.user_service.v1.UserService"
        ]
      }
    }
  },
  "definitions": {
    "proto.user_service.v1.CreateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "proto.user_service.v1.GetUserByIDResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/proto.user_service.v1.User"
        }
      }
    },
    "proto.user_service.v1.GetUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/proto.user_service.v1.User"
          }
        }
      }
    },
    "proto.user_service.v1.MutationResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "proto.user_service.v1.User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "proto.user_service.v1.UserService.DeleteUserBody": {
      "type": "object"
    },
    "proto.user_service.v1.UserService.UpdateUserBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    }
  }
}